Move publish_sns_message function to common lib
