+++ lambdas/service_scheduler/sns_utils.py
@@ -1,0 +1,1 @@
+../common/sns_utils.py
